backend/pyproject.toml
[tool.poetry]
name = "article-gpt"
version = "0.1.0"
description = ""
authors = ["Andre du Plessis <andredp@aleios.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.30.0"
requests = "^2.31.0"
bs4 = "^0.0.1"
langchain = "^0.1.0"
langchain-community = "^0.0.11"
langchain-openai = "^0.0.2"
langchain-core = "^0.1.10"
python-dotenv = "1.0.0"
chromadb = "^0.4.22"
supabase = "^2.3.3"
pydantic-settings = "^2.1.0"
fake-useragent = "^1.4.0"
selenium = "^4.17.2"
pdfminer-six = "^20231228"
boto3 = "^1.34.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
